A lot of permissions-related stuff. Here goes:
groups, devices, users, and permissions have resource objects. They can be secured with OSF (Octoblu Security Framework)?.
We may want to secure who can even *see* the permissions on a device. In fact, we probably do.

Made a sort of mixin-factory that does all the heavy lifting of adding resource properties to a schema. It's a one-liner now to add resource functionality to new entities!

 *All* resource concepts now go through resourceIds, rather than just random properties. the Resource.makeResource() function will map an existing property to a resource property.

 This includes members in groups: they are resourceIds now. The reason we had to do this is because a uuid won't tell you where to look for the entity - if members were uuids, we would never be able to retrieve them. See me griping about Mongoose below if you want to know why we need the .resourceId property on resources.

resource stuff is emulated on existing user models, but would persist if the user was then saved. The user database can be migrated by simply loading all the users, then saving them again. Until this happens, USERS & DEVICES CANNOT BE POPULATED IN GROUPS. So just add new users to fix this.

Creating a group will cause the source, target, and permissions to be created, and the resource data (parent, owner, uuid, type) of all entities are set correctly! The members can also be updated if you have a valid resourceId for them.

**** complaints ****
You may notice that whenever we set a reference to a resource, we annoyingly have to call .resourceId. E.g. resource.resourceId, owner.resourceId, etc. But not for reading the properties. Why?

Because Mongoose stores subdocuments in a proxied object ("model") that can not be directly assigned to an object literal. Well, it normally could, but Mongoose can't seem to tell that the proxied object has the same exact properties as the object literal, and so it fails validation every single time, silently.